langchain/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
dataclasses-json==0.6.6
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
frozenlist==1.4.1
greenlet==3.0.3
grpcio==1.64.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-path==0.3.2
jsonschema-specifications==2023.12.1
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langsmith==0.1.60
lazy-object-proxy==1.10.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
more-itertools==10.2.0
motor==3.4.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.30.1
openapi-core==0.19.1
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
orjson==3.10.3
packaging==23.2
parse==1.20.1
pathable==0.4.3
prance==23.6.21.0
pybars4==0.9.13
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
PyMeta3==0.5.1
pymongo==4.7.2
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.31.1
regex==2023.12.25
requests==2.32.1
rfc3339-validator==0.1.4
rich==13.7.1
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scipy==1.13.0
semantic-kernel==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4
